+++ src/math/iir.c
@@ -7,8 +7,8 @@
 //         Keyon Jie <yang.jie@linux.intel.com>
 
 #include <sof/common.h>
-#include <sof/audio/eq_iir/iir.h>
 #include <sof/audio/format.h>
+#include <sof/math/iir_df2t.h>
 #include <user/eq.h>
 #include <errno.h>
 #include <stddef.h>
