+++ test/cmocka/src/audio/eq_iir/CMakeLists.txt
@@ -13,7 +13,7 @@ add_compile_options(-DUNIT_TEST)
 
 add_library(audio_for_eq_iir STATIC
 	${PROJECT_SOURCE_DIR}/src/audio/eq_iir/eq_iir.c
-	${PROJECT_SOURCE_DIR}/src/audio/eq_iir/iir.c
+	${PROJECT_SOURCE_DIR}/src/math/iir.c
         ${PROJECT_SOURCE_DIR}/src/math/iir_df2t_generic.c
         ${PROJECT_SOURCE_DIR}/src/math/iir_df2t_hifi3.c
 	${PROJECT_SOURCE_DIR}/src/audio/buffer.c
